Uninstall description + launch gpx.exe dir
The nuspec description gets put in the Windows control panel as the
program name. Made it shorter to be more explicable, perhaps we should
go shorter yet.

In anticipation of not necessarily having gpx.exe on the path (plus
controlling for multiple versions of gpx.exe), launch gpx.exe from the
same folder as gpxui.exe.