Let's keep a copy of the .ini's in the parent
folder.  This allows us to easily copy them down again when we install a
new version because squirrel puts the new version in a new folder.
Probably we should just teach gpx.exe where to find the gpx.ini in the
unversioned appdata folder, but for now we'll leave it as is so that
people can use their own version of gpx.exe (would that happen?)